+++ pages/forgot-password.tsx
@@ -1,5 +1,5 @@
 import Body from "@/components/layout/Body";
-import TextInput from "@/components/placeholderfc/TextInput";
+import TextInput from "@/components/formControllers/TextInput";
 import { PAGES } from "@/helpers/navigation";
 import useAuth from "@/hooks/useAuth";
 import {
